Modified Image Class to use 2D array, more Helpers
The Image class now stores and iterates over the __image array as a 2D
array. This makes calculating the data much more convenient internally as
well as means the input data doesn't need to be transformed into a 1D
array before creating an Image.
Refactored the Image class some more to make the code as consistent as
possible.
Added the Perceptron class to Helpers. It may still have some issues.
Modified the test cases in main that sanity check the Image class so they
generate 2D arrays now.
Started working with the mnist data using numpy.